added Exception

+++ src/main/java/com/practice/EmployeeManagement/execption/EmployeeNotFound.java
@@ -1,7 +1,7 @@
 package com.practice.EmployeeManagement.execption;
 
 public class EmployeeNotFound extends Exception {
-    public EmployeeNotFound(Long id) {
+    public EmployeeNotFound(int id) {
         super(String.format("Employee Not Found of id %d ", id));
     }
 }
